requirements.txt
# django>=4.0.0,<4.1.0
djangorestframework
pyyaml
requests
django-cors-headers
Pillow>=9.0
PyJWT==2.1.0
pytz==2021.1
asgiref==3.4.1
sqlparse==0.4.1